created enemy spawner prefab, added the spawner into the sample scene.
Created an object for an enemy spawner. spawner is a script that has a minimum time and macimum time and will instantiate a prefab. Can place the spawner prefab at bottom 4 corners of tub and have the germs float up. Can even choose which enemy prefab is spawned so we can implement the large germ.